refactor: rename SummaryProjectLogo
