+++ frontend/src/lib/components/summary/SummaryProjectLogo.svelte
