+++ frontend/src/tests/lib/components/summary/SummaryProjectLogo.spec.ts
@@ -2,7 +2,7 @@
  * @jest-environment jsdom
  */
 
-import ProjectLogo from "$lib/components/summary/ProjectLogo.svelte";
+import SummaryProjectLogo from "$lib/components/summary/SummaryProjectLogo.svelte";
 import { OWN_CANISTER_ID_TEXT } from "$lib/constants/canister-ids.constants";
 import { IC_LOGO } from "$lib/constants/icp.constants";
 import { snsProjectSelectedStore } from "$lib/derived/selected-project.derived";
@@ -12,12 +12,12 @@ import { mockStoreSubscribe } from "../../../mocks/commont.mock";
 import { mockSnsFullProject } from "../../../mocks/sns-projects.mock";
 import { mockSnsCanisterIdText } from "../../../mocks/sns.api.mock";
 
-describe("ProjectLogo", () => {
+describe("SummaryProjectLogo", () => {
   describe("nns", () => {
     beforeEach(() => page.mock({ data: { universe: OWN_CANISTER_ID_TEXT } }));
 
     it("should render ic logo", () => {
-      const { getByTestId } = render(ProjectLogo);
+      const { getByTestId } = render(SummaryProjectLogo);
 
       expect(getByTestId("logo")?.getAttribute("src")).toEqual(IC_LOGO);
     });
@@ -33,7 +33,7 @@ describe("ProjectLogo", () => {
     beforeEach(() => page.mock({ data: { universe: mockSnsCanisterIdText } }));
 
     it("should render project logo", () => {
-      const { getByTestId } = render(ProjectLogo);
+      const { getByTestId } = render(SummaryProjectLogo);
 
       expect(getByTestId("logo")?.getAttribute("src")).toEqual(
         mockSnsFullProject.summary.metadata.logo
